Changed the FileManager algorithm that copies files from the cache / the exclusion list to the target directory to only copy if the file doesn't exist or the new file timestamp is more recent.

+++ src/scweaver/FileManager.cs
@@ -131,7 +131,9 @@ namespace Weaver {
         void MirrorFilesToCopy() {
             foreach (var item in filesToCopy) {
                 var target = Path.GetFullPath(Path.Combine(TargetDirectory, Path.GetFileName(item)));
-                File.Copy(item, target, true);
+                if (!File.Exists(target) || File.GetLastWriteTime(target) < File.GetLastWriteTime(item)) {
+                    File.Copy(item, target, true);
+                }
             }
         }
